Added movies app screenshots to Projects page
To accomplish this I added two new components, Movies and MoviesPictures. The layout is exactly the same as Marfare and MarfarePictures. I changed the marfare container id's to generic classNames and applied those classNames to the Movies containers. The github link currently links to the marfare repo. I will update it as soon as I can make the movies_app repo public.